+++ src/components/base/b-tree/test/helpers.ts
@@ -6,7 +6,7 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import type { BrowserContext, ElementHandle, JSHandle, Page } from 'playwright';
+import type { ElementHandle, JSHandle, Page } from 'playwright';
 
 import test from 'tests/config/unit/test';
 
@@ -14,17 +14,16 @@ import DOM from 'tests/helpers/dom';
 import Component from 'tests/helpers/component';
 
 import type bTree from 'components/base/b-tree/b-tree';
-import type { Item } from 'components/base/b-tree/interface';
-
-interface CheckOptionTreeCtx {
-	target: JSHandle<bTree>;
-	queue?:Array<Promise<void>>;
-	level?: number;
-	foldSelector?: string;
-}
+import type { Item } from 'components/base/b-tree/b-tree';
 
 /**
- * Returns the rendered `b-tree` component
+ * Returns a selector for the passed bTree element
+ * @param elName
+ */
+export const createTreeSelector = DOM.elNameSelectorGenerator('b-tree');
+
+/**
+ * Returns the rendered bTree component
  *
  * @param page
  * @param opts
@@ -33,7 +32,11 @@ export async function renderTree(
 	page: Page,
 	opts: Partial<{items: Item[]} & RenderComponentsVnodeParams> = {}
 ): Promise<JSHandle<bTree>> {
-	const {items, attrs, children} = opts;
+	const {
+		items,
+		attrs,
+		children
+	} = opts;
 
 	return Component.createComponent(page, 'b-tree', {
 		attrs: {
@@ -48,75 +51,45 @@ export async function renderTree(
 }
 
 /**
- * Returns the `folded` class modifier for the given value
- *
- * @param target
- * @param value
- */
-export function getFoldedClass(target: JSHandle<bTree>, value: boolean = true): Promise<string> {
-	return target.evaluate(
-		(ctx, v) => ctx.unsafe.block!.getFullElementName('node', 'folded', v),
-		value
-	);
-}
-
-/**
- * Iterates over all tree items depth-first and unfolds closed items by clicking them.
- * Also, it checks that all items have correct `data-level` attribute and `folded` class modifier.
- *
- * @param page
+ * Returns the total number of tree items based on the value of the component's prop
  * @param items
- * @param param2
  */
-export function checkOptionTree(
-	page: Page,
-	items: Item[],
-	{target, queue = [], level = 0, foldSelector}: CheckOptionTreeCtx
-): Array<Promise<void>> {
-	items.forEach((item) => {
-		const isBranch = Object.isArray(item.children);
+export function getItemsCount(items: Item[]): number {
+	let count = 0;
 
-		const promise = (async () => {
-			const
-				id = await target.evaluate((ctx, value) => `${ctx.unsafe.values.getIndex(value)}`, item.value),
-				elementSelector = `[data-id="${id}"]`,
-				element = await page.waitForSelector(elementSelector, {state: 'attached'});
+	items.forEach(({children}) => {
+		count++;
 
-			const
-				isFolded = item.folded ?? await target.evaluate((ctx) => ctx.folded),
-				foldedClass = await getFoldedClass(
-					target,
-					isFolded
-				);
-
-			await test.expect(element.getAttribute('data-level')).toBeResolvedTo(String(level));
-
-			if (isBranch) {
-				const
-					selector = foldSelector ?? DOM.elNameSelectorGenerator('b-tree', 'fold'),
-					fold = await element.waitForSelector(selector, {state: 'attached'});
-
-				await test.expect(page.locator(elementSelector)).toHaveClass(new RegExp(foldedClass));
-
-				if (isFolded) {
-					await fold.click();
-				}
-			}
-		})();
-
-		queue.push(promise);
-
-		if (isBranch) {
-			checkOptionTree(page, item.children ?? [], {
-				level: level + 1,
-				target,
-				queue,
-				foldSelector
-			});
+		if (children != null) {
+			count += getItemsCount(children);
 		}
 	});
 
-	return queue;
+	return count;
+}
+
+/**
+ * Returns the number of rendered tree items
+ * @param tree
+ */
+export function getRenderedNodesCount(tree: JSHandle<bTree>): Promise<number> {
+	return tree.evaluate((ctx) => {
+		const nodes = ctx.$el!.querySelectorAll(`.${ctx.provide.fullElementName('node')}`);
+		return nodes.length;
+	});
+}
+
+/**
+ * Returns a promise that will not resolve until the tree has a specified number of rendered items
+ *
+ * @param tree
+ * @param number
+ */
+export function waitForNumberOfNodes(tree: JSHandle<bTree>, number: number): Promise<boolean> {
+	return tree.evaluate(({unsafe}, number) => unsafe.async.wait(() => {
+		const nodes = unsafe.$el!.querySelectorAll(`.${unsafe.provide.fullElementName('node')}`);
+		return nodes.length === number;
+	}), number);
 }
 
 /**
@@ -126,7 +99,7 @@ export function checkOptionTree(
  * @param target
  * @param value
  */
-export async function waitForItem(
+export async function waitForItemWithValue(
 	page: Page,
 	target: JSHandle<bTree>,
 	value: unknown
@@ -136,13 +109,13 @@ export async function waitForItem(
 }
 
 /**
- * Waits for rendering of items with given values
+ * Waits for rendering of items with the given values
  *
  * @param page
  * @param target
  * @param values
  */
-export async function waitForItems(
+export async function waitForItemsWithValues(
 	page: Page,
 	target: JSHandle<bTree>,
 	values: Iterable<unknown>
@@ -156,96 +129,15 @@ export async function waitForItems(
 }
 
 /**
- * Provides an API to intercept and mock response for the b-tree request
- * @param pageOrContext
- */
-export function interceptTreeRequest(
-	pageOrContext: Page | BrowserContext
-): Promise<void> {
-	return pageOrContext.route(/api/, async (route) => route.fulfill({
-		status: 200,
-		contentType: 'application/json',
-		body: JSON.stringify([
-			{value: 'foo_0_0'},
-			{
-				value: 'foo_0_1',
-				children: [
-					{value: 'foo_1_0'},
-					{value: 'foo_1_1'},
-
-					{
-						value: 'foo_1_2',
-						children: [{value: 'foo_2_0'}]
-					},
-
-					{value: 'foo_1_3'},
-					{value: 'foo_1_4'},
-					{value: 'foo_1_5'}
-				]
-			},
-			{value: 'foo_0_2'},
-			{value: 'foo_0_3'},
-			{value: 'foo_0_4'},
-			{value: 'foo_0_5'},
-			{value: 'foo_0_6'}
-		])
-	}));
-}
-
-/**
- * Returns a selector for the passed element
- * @param elName
- */
-export const createTreeSelector = DOM.elNameSelectorGenerator('b-tree');
-
-/**
- * Creates a function to test if nodes have given modifier classes
+ * Creates a function to test if nodes have the given modifier classes
  * @param modName
  */
-export function createTestModIs(modName: string) {
+export function createExpectMod(modName: string) {
 	return async (
 		status: boolean,
 		nodes: Array<ElementHandle<HTMLElement | SVGElement>>
 	): Promise<void> => {
 		const classes = await Promise.all(nodes.map((node) => node.getAttribute('class')));
-
-		test.expect(classes.every((x) => x?.includes(status ? `${modName}_true` : `${modName}_false`)))
-			.toBeTruthy();
+		test.expect(classes.every((x) => x?.includes(status ? `${modName}_true` : `${modName}_false`))).toBeTruthy();
 	};
-}
-
-/**
- * Checks if page has expected count of b-checkbox elements
- *
- * @param page
- * @param expectedCount
- */
-export async function waitForCheckboxCount(page: Page, expectedCount: number): Promise<void> {
-	await test.expect(page.locator('.b-checkbox')).toHaveCount(expectedCount);
-}
-
-/**
- * Returns default items for tests
- */
-export function getDefaultItems(): Item[] {
-	return [
-		{value: 'bar'},
-
-		{
-			value: 'foo',
-			children: [
-				{value: 'foo_1'},
-				{value: 'foo_2'},
-
-				{
-					value: 'foo_3',
-					children: [{value: 'foo_3_1'}]
-				},
-
-				{value: 'foo_4'},
-				{value: 'foo_5'},
-				{value: 'foo_6'}
-			].map((item) => ({...item, label: item.value}))
-		}
-	].map((item) => ({...item, label: item.value}));
 }