+++ src/components/base/b-bottom-slide/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.55 (2024-02-08)
+
+#### :bug: Bug Fix
+
+* Fixed move to the closest step in `heightMode = content`
+
 ## v4.0.0-alpha.1 (2022-12-14)
 
 #### :boom: Breaking Change